batman

+++ src/main/java/com/example/h09_nop_relexsun_spring_homework002/controller/StudentController.java
@@ -1,5 +1,6 @@
 package com.example.h09_nop_relexsun_spring_homework002.controller;
 
+import com.example.h09_nop_relexsun_spring_homework002.exception.NotFoundException;
 import com.example.h09_nop_relexsun_spring_homework002.model.DTO.request.StudentRequest;
 import com.example.h09_nop_relexsun_spring_homework002.model.DTO.response.APIResponse;
 import com.example.h09_nop_relexsun_spring_homework002.model.entity.Student;
@@ -27,14 +28,14 @@ public class StudentController {
     }
 
     @PostMapping
-    public ResponseEntity<APIResponse<Student>> createStudent(@Valid @RequestBody StudentRequest request) {
+    public ResponseEntity<APIResponse<Student>> createStudent(@RequestBody @Valid StudentRequest request) {
         Student student = this.studentService.createStudent(request);
         APIResponse<Student> response = APIResponse.<Student>builder().message("success").payload(student).status(HttpStatus.OK).timestamp(Instant.now()).build();
-        return ResponseEntity.status(HttpStatus.OK).body(response);
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
     @PutMapping("/{student-id}")
-    public ResponseEntity<APIResponse<Student>> updateStudent(@PathVariable("student-id") Long id, @Valid @RequestBody StudentRequest request) {
+    public ResponseEntity<APIResponse<Student>> updateStudent(@PathVariable("student-id") Long id, @RequestBody @Valid StudentRequest request) {
         Student student = this.studentService.updateStudent(id, request);
         APIResponse<Student> response = APIResponse.<Student>builder().message("success").payload(student).status(HttpStatus.OK).timestamp(Instant.now()).build();
         return ResponseEntity.status(HttpStatus.OK).body(response);
@@ -43,12 +44,15 @@ public class StudentController {
     @GetMapping("/{student-id}")
     public ResponseEntity<APIResponse<Student>> getStudentById(@PathVariable("student-id") Long id) {
         Student student = this.studentService.getStudentById(id);
+        if(student == null) throw new NotFoundException("Student not found");
         APIResponse<Student> response = APIResponse.<Student>builder().message("success").payload(student).status(HttpStatus.OK).timestamp(Instant.now()).build();
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 
     @DeleteMapping("/{student-id}")
     public ResponseEntity<APIResponse<Student>> deleteStudentById(@PathVariable("student-id") Long id) {
+        Student student = this.studentService.getStudentById(id);
+        if(student == null) throw new NotFoundException("Student not found");
         this.studentService.deleteStudentById(id);
         APIResponse<Student> response = APIResponse.<Student>builder().message("Student deleted successfully").payload(null).status(HttpStatus.OK).timestamp(Instant.now()).build();
         return ResponseEntity.status(HttpStatus.OK).body(response);
